Extracted bot config to separate module + minor cleanups
